Avoid NoMethodError when someone hangs up
When we handle a hangup, we also deal with the Agent's call. This handling was added specifically for the bridge strategy, but this strategy depends on several items implemented in an application outside of ES, and is incomplete. It does, however, introduce the requirement that `ElectricSlide::Agent#call` exist, but not that it ever take a value.

This is sufficient to get outbound calling to agents working, but does not complete the bridge strategy, which should be completed in ES with a re-evaluation of why tests pass with such bugs present.