Adding support for queue class pre-validation
Actors which fail to start will crash the supervisor. This commit
provides Electric Slide with a way to check if a queue can be safely
created before it is started. This allows Electric Slide to skip trying
to start the queue if queue creation will raise an exception.

A `::valid_with?` class method, which is given a set of queue
attributes, is implemented in `ElectricSlide::CallQueue`.