Allowing `#add_agent` to finish without waiting to establish connections
Once `ElectricSlide::CallQueue#add_agent` is done preparing and adding
the agent into the queue, it immediately checks if the agent can be
connected to a waiting call before returning.

This commit changes that behavior by returning from `#add_agent`
immediately after the agent is processed into the queue. The connection
attempt is still requested, but asynchronously.