Conditionally clear the agents call object when returning the agent
We use the agent's state in an Electric Slide Call Queue to build
dashboards that show the agents who are signed in and what calls they
are on. The dashboards all work fine, except when a call queued and
wait for an agent. In these cases, often, we lose the information we
need to know what call the agent is on.

This bug is easiest to replicate with one agent signed in and two
calls get placed simultaneously. One of the two calls will queue, and
the other will go straight to the available agent. The dashboard will
work fine at this point, and we can report what call the agent is
on. When the call ends, the queued call will also go to this agent,
but the dashboard will not have the information needed to report what
call the agent is on.

When we connect an agent to a call, we register a call end
handler. The call end handler will return the agent to make them
available for calls and clear out the agents all. When we return the
agent, we also check to see if any calls are waiting in the queue. If
calls are waiting, we connect an available agent to the next queued
call. Sometimes the next available agent is the same as the agent
whose call is ending. When the agent connects with the queued call, we
update the agent's call object. At this point, we have the data we
need to report correctly on what call the agent is on. But then the
code moves on to the next step in the call end hander, and we clear
out the agent's call and loos the ability to report what cal the agent
is on.

To fix this, we now conditionally clear the agent's call. Before
checking to see if any are waiting in the queue and connecting the
next call, we save away the agent's current call object. If the call
object is the same as before, we clear out the agent's call object,
but if it has changed, we skip clearing out the agent's call object.